+++ apps/be/src/main/java/com/capstone/logue/anal/dto/fastapi/response/WarningInfo.java
@@ -1,4 +1,4 @@
-package com.capstone.logue.anal.dto.fastapi;
+package com.capstone.logue.anal.dto.fastapi.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
